Function cap_rand::std_rng_from_entropy

pub fn std_rng_from_entropy(_: AmbientAuthority) -> StdRng

Retrieve the standard random number generator, seeded by the system.

This corresponds to rand::rngs::StdRng::from_entropy.

Ambient Authority

This function makes use of ambient authority to access the platform entropy source.