Module cap_rand::prelude

Convenience re-export of common members.

This corresponds to rand::prelude.

Re-exports

pub use crate::rngs::CapRng;

pub use crate::random;

pub use crate::thread_rng;

Structs

StdRng

The standard RNG. The PRNG algorithm in StdRng is chosen to be efficient on the current platform, to be statistically strong and unpredictable (meaning a cryptographically secure PRNG).

Traits

CryptoRng

A marker trait used to indicate that an RngCore or BlockRngCore implementation is supposed to be cryptographically secure.

Distribution

Types (distributions) that can be used to create a random instance of T.

IteratorRandom

Extension trait on iterators, providing random sampling methods.

Rng

An automatically-implemented extension trait on RngCore providing high-level generic methods for sampling values and other convenience methods.

RngCore

The core of a random number generator.

SeedableRng

A random number generator that can be explicitly seeded.

SliceRandom

Extension trait on slices, providing random mutation and sampling methods.